# backend/swaggerDoc.yaml
openapi: 3.0.3
info:
  title: Disaster Management API
  version: 1.0.0
servers:
  - url: http://localhost:5555/api
    description: development server
  - url: https://cdrp-1.onrender.com/api
    description: production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CharityAd:
      type: object
      required: [charity, title, description, duration, expiresAt]
      properties:
        _id: { type: string }
        charity: { type: string }
        title: { type: string, minLength: 4, maxLength: 50 }
        image: { type: string, format: uri }
        cloudinaryId: { type: string }
        description: { type: string, minLength: 4, maxLength: 1000 }
        status: { type: string, enum: [open, closed], default: open }
        volunteers: { type: array, items: { type: string } }
        duration: { type: number }
        expiresAt: { type: string, format: date-time }
        categories: { type: array, items: { type: string } }
        requirements:
          {
            type: object,
            properties:
              {
                location: { type: string },
                skills: { type: array, items: { type: string } },
              },
          }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /charityad:
    get:
      tags: [CharityAds]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 5 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          {
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/CharityAd" },
                      },
                  },
              },
          }
    post:
      tags: [CharityAds]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                charity: { type: string }
                title: { type: string }
                description: { type: string }
                duration: { type: number }
                categories: { type: array, items: { type: string } }
                requirements:
                  {
                    type: object,
                    properties:
                      {
                        location: { type: string },
                        skills: { type: array, items: { type: string } },
                      },
                  }
                image: { type: string, format: binary }
      responses:
        "201":
          {
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CharityAd" } },
              },
          }
    delete:
      tags: [CharityAds]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { content: { application/json: {} } }
...